pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easyslice"
version = "0.1.0"
description = "Turn longform videos into short story clips with captions."
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
  "pydantic>=2.6",
]

# For a pinned environment snapshot, see requirements.txt.

[project.optional-dependencies]
media = [
  "yt-dlp>=2024.8.6",
  "openai-whisper>=20231117",
  "moviepy>=2.2.1",
  "pillow>=10.0",
  "numpy>=1.24",
]

dev = [
  "pytest>=7.4",
]

[project.scripts]
easyslice = "easyslice.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
easyslice = ["py.typed"]

[tool.pytest.ini_options]
addopts = "-q"
